+++ __init__.py
@@ -1,1 +1,1 @@
-from api import *
+from . import *
