+++ network/udp_api.py
@@ -1,8 +1,6 @@
 import socket
-import api
+from default import api
 import netifaces
-import threading
-import os
 import time
